% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{Add a row to one of the tables in the "database"}
\usage{
add_row(data, table = "events")
}
\arguments{
\item{data}{A dataframe}

\item{table}{One of "credentials", "events", "offers", "transactions" or "users"}
}
\value{
A row added to the corresponding google sheet
}
\description{
Add a row to the google sheet of
}

