% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user.R
\name{add_user}
\alias{add_user}
\title{Add a user to the users table}
\usage{
add_user(data = NULL, user_id = 1, name = "John Doe",
  email = "joebrew@gmail.com", username = "username",
  password = "password", amount = 100, amount_invested = 0,
  amount_offered = 0)
}
\arguments{
\item{data}{A dataframe}

\item{user_id}{A user id number}

\item{name}{A name}

\item{email}{An email}

\item{username}{A username}

\item{password}{A password}

\item{amount}{An amount}

\item{amount_invested}{An amount invested}

\item{amount_offered}{An amount offered}
}
\value{
A row added to the corresponding google sheet
}
\description{
Add a row to the google sheet of
}

