% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spread.R
\name{get_spread}
\alias{get_spread}
\title{Get spread}
\usage{
get_spread(offers = NULL, events = NULL, plot = TRUE, eid = NULL)
}
\arguments{
\item{offers}{An offers table}

\item{events}{An events table}

\item{plot}{Whether to plot or not}

\item{eid}{Optional, an event id to filter by}
}
\value{
A row added to the corresponding google sheet
}
\description{
Get the current spread of all offers
}

