% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invite_participants.R
\name{invite_participants}
\alias{invite_participants}
\title{Invite participants to participate in the market}
\usage{
invite_participants()
}
\arguments{
\item{data}{A dataframe}

\item{table}{One of "credentials", "events", "offers", "transactions" or "users"}
}
\value{
A row added to the corresponding google sheet
}
\description{
Add a row to the google sheet of
}

