% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_data.R
\name{load_all_data}
\alias{load_all_data}
\title{Read all tables from the "database"}
\usage{
load_all_data()
}
\value{
Several tables read into memory
}
\description{
Read all the google sheets for the malaria prediction market project
}

