% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_table.R
\name{load_table}
\alias{load_table}
\title{Read an entire table from the "database"}
\usage{
load_table(table = "events")
}
\arguments{
\item{table}{One of "credentials", "events", "offers", "transactions" or "users"}

\item{data}{A dataframe}
}
\value{
A row added to the corresponding google sheet
}
\description{
Read an entire google sheet
}

