% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_random_offers.R
\name{make_random_offers}
\alias{make_random_offers}
\title{Make random offers from random users for random event ids}
\usage{
make_random_offers(n = 20)
}
\arguments{
\item{n}{The number of random offers to make}
}
\value{
A row added to the offers table
}
\description{
Make random offers from random users for random events. This is used for testing only.
}

