% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transactions.R
\name{plot_transactions}
\alias{plot_transactions}
\title{Plot transactions for an event}
\usage{
plot_transactions(event_id, transactions = NULL)
}
\arguments{
\item{event_id}{an event id}

\item{transactions}{a transactions table}
}
\value{
A row added to the corresponding google sheet
}
\description{
Plot transactions for a certain event
}

