% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_users.R
\name{set_up_users}
\alias{set_up_users}
\title{Set up the users table}
\usage{
set_up_users()
}
\value{
The users spreadsheet will be populated
}
\description{
Populate the users table. Should only be run once
}

