% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_user.R, R/invite_users.R
\name{add_user}
\alias{add_user}
\title{Add a user to the users table}
\usage{
add_user(users, n = NULL, password = "")

add_user(users, n = NULL, password = "")
}
\arguments{
\item{n}{The number of users to invite}

\item{password}{A password}

\item{data}{A dataframe}

\item{user_id}{A user id number}

\item{name}{A name}

\item{email}{An email}

\item{username}{A username}

\item{amount}{An amount}
}
\value{
A row added to the corresponding google sheet

Users will be emailed
}
\description{
Add a row to the google sheet of

Invite users to participate
}

